requirements.txt
discord.py==2.2.2
requests==2.28.2
umongo==3.1.0
redis==4.5.5
aioredis==2.0.1
pandas==2.0.1